⚓️:: [#153] domain model 추가
