+++ Projects/Domain/Sources/Model/Home/StepModel.swift
@@ -1,0 +1,17 @@
+//
+//  StepModel.swift
+//  Domain
+//
+//  Created by 박준하 on 1/30/24.
+//  Copyright © 2024 MaeumGaGym-iOS. All rights reserved.
+//
+
+import UIKit
+
+public struct StepModel {
+    public var stepCount: Int
+    
+    public init(stepCount: Int) {
+        self.stepCount = stepCount
+    }
+}
