--- Projects/Features/HomeFeature/Sources/HomeScene/TableViewCell/Extra/ExtrasModel.swift
@@ -1,7 +1,0 @@
-import UIKit
-
-public struct ExtrasModel {
-    public var image: UIImage
-    public var titleName: String
-    public var description: String
-}
