--- Projects/Features/HomeFeature/Sources/HomeScene/TableViewCell/Routine/RoutineModel.swift
@@ -1,7 +1,0 @@
-import UIKit
-
-public struct RoutineModel {
-    public var exercise: String
-    public var sets: Int
-    public var reps: Int
-}
